ロック画面コントロールを長押ししてもアプリが開かない問題を修正 (#6572)
* ロック画面コントロールを長押ししてもアプリが開かない問題を修正


* Universal Link未設定のためOpenURLIntent併用を取りやめてopenAppWhenRunに一本化


---------
